backend/pyproject.toml
[build-system]
requires = ["setuptools>=80.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "m3u8-flow"
version = "0.1.0"
description = "M3U8 Download"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiolimiter>=1.2.1",
    "fastapi[standard]>=0.135.1",
    "gunicorn>=25.3.0",
    "httpx-retries>=0.4.6",
    "httpx[socks,http2]>=0.28.1",
    "loguru>=0.7.3",
    "m3u8>=6.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "sse-starlette>=3.3.3",
    "tortoise-orm>=1.1.7",
    "tzdata>=2025.2",
]

[dependency-groups]
dev = ["asgi-lifespan>=2.1.0", "pytest>=9.0.2", "pytest-asyncio>=1.3.0"]

[tool.setuptools]
package-dir = { "app" = "app" }

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = ["-vv", "-s"]
# API 测试请用 httpx.AsyncClient(ASGITransport)，勿用 Starlette TestClient（anyio blocking portal 与 session 事件循环 teardown 交织时易导致 pytest 进程不退出）
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
pythonpath = ["."]
markers = [
    "integration: integration tests (often require network or external resources)",
]

# Tortoise CLI 读取：键名必须为 tortoise_orm（见 tortoise.cli.utils.tortoise_orm_config）
# 值为「模块路径.变量名」，需指向含 TORTOISE_ORM 字典的模块
[tool.tortoise]
tortoise_orm = "app.core.config.TORTOISE_ORM"
